weather_dashboard/requirements.txt
Flask
pandas
psycopg2-binary